Split Redhat and Ubuntu variants of seed files
Without this change, the seed files for Unbound and Consul were becoming
messy in order to deal with both Redhat and Ubuntu installations.  Recent
refactoring of the deployment code allows them to be split into separate
files and the right one pulled in automatically.  This change takes
advantage of that to clean up the seed files.