Move desployer.conf so it's not used by default
The included desployer.conf is an example, which gets in the way of finding
a host-specific version in ~/.desployer.conf or /etc/desployer.conf.  This
change renames it to desployer.conf.example so that the host-specific
versions can be used when running desployer from a clone of the git repo.