Add capability to use raw disks rather than qcow2 snapshots
Without this change, qcow2 snapshots are used to provision the root disk
from an image and additional disks are provisioned using the qcow2 format.
While qcow2 is excellent for quickly spinning up VMs to run functional tests
against, it doesn't perform as well as raw disk thus isn't ideal for VMs
being deployed for production use.

This change adds the capability to use raw disks - which take a little
longer to provision, but perform better once provisioned - and makes it the
default.  Using qcow2 snapshots is still supported but now requires the
--snapshot command line option be passed to the "up" subcommand, e.g.:

  $ desployer up --snapshot

Note that the --snapshot option only applies to the root disk.  This patch
switches the deployment of perpetual disks from qcow2 format to raw (and
unlike the root disk, there is no flag to use the old behaviour).