Avoid the -H option for cloud-localds
Without this change, cloud-localds uses the -H option rather than generating
the metadata itself.  This is a problem because the -H option is not
available in cloud-localds from RHEL6's cloud-utils-0.27-10.el6.x86_64
package.

Note that the manually generated metadata this patch adds is identical to
the metadata generated by using the -H option (when the -H option exists, as
it does in Ubuntu 14.04).